Add files via upload
This is the register page and login page of the staff uog forum. This is as far I can get without the MySQL database to test with it. Will update the C# pages in the register page and login page once the MySQL database is developed. Couldn't upload the register page and login page together since Github does not allow file uploads that are over 25MB.